Move filter adapters to their own files (#73720)
This refactors a part of the `filters` aggregator into multiple files so
its a little easier to read. We ended up with more subclasses then we
expected at first.
